requirements.txt
websockets>=10.0
pyzmq>=22.0
protobuf>=3.19
